Move all multi-seller and single-seller functions to its own folder
