+++ packages/explorable-explanations/src/protectedAudience/modules/auctions/index.js
@@ -22,6 +22,12 @@ import config from '../../config';
 import * as utils from '../../utils';
 import { Box, ProgressLine, Branches, RippleEffect } from '../../components';
 import bubbles from '../bubbles';
+import setUpSingleSellerFirstSSPTagFlow from './single-seller/setupFirstSSPTagFlow';
+import setUpMultiSellerFirstSSPTagFlow from './multi-seller/setUpFirstSSPTagFlow';
+import setUpPublisherAdServerFlow from './multi-seller/setUpPublisherAdServerFlow';
+import setUpComponentAuction from './multi-seller/setUpComponentAuction';
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
+import setUpRunadAuction from './setUpRunadAuction';
 
 /**
  * @module Auction
@@ -61,14 +67,14 @@ auction.setUp = (index) => {
   auction.setupBranches(index);
 
   if (app.isMultiSeller) {
-    auction.setUpMultiSellerFirstSSPTagFlow();
-    auction.setUpPublisherAdServerFlow();
-    auction.componentAuction();
+    setUpMultiSellerFirstSSPTagFlow(auction);
+    setUpPublisherAdServerFlow(auction);
+    setUpComponentAuction(auction);
   } else {
-    auction.setUpSingleSellerFirstSSPTagFlow();
+    setUpSingleSellerFirstSSPTagFlow(auction);
   }
 
-  // auction.setUpRunadAuction();
+  // setUpRunadAuction(auction);
   // auction.setupShowWinningAd();
 
   app.auction.auctions.push(auction.steps);
@@ -140,641 +146,6 @@ auction.setupBranches = (index) => {
     callBack: (returnValue) => {
       app.auction.nextTipCoordinates = returnValue;
     },
-  });
-};
-
-auction.setUpSingleSellerFirstSSPTagFlow = () => {
-  const { box, colors } = config.flow;
-
-  auction.steps.push({
-    component: ProgressLine,
-    props: {
-      direction: 'down',
-      x1: () => app.auction.nextTipCoordinates?.x,
-      y1: () => app.auction.nextTipCoordinates?.y + 40,
-      noArrow: true,
-    },
-    callBack: (returnValue) => {
-      app.auction.nextTipCoordinates = returnValue;
-    },
-  });
-
-  auction.steps.push({
-    component: Box,
-    props: {
-      title: 'SSP Tag',
-      x: () => app.auction.nextTipCoordinates?.x - box.width / 2,
-      y: () => app.auction.nextTipCoordinates?.y,
-    },
-    delay: 1000,
-    callBack: (returnValue) => {
-      app.auction.nextTipCoordinates = returnValue.down;
-    },
-  });
-
-  auction.steps.push({
-    component: ProgressLine,
-    props: {
-      direction: 'down',
-      x1: () => app.auction.nextTipCoordinates?.x,
-      y1: () => app.auction.nextTipCoordinates?.y + 40,
-    },
-    callBack: (returnValue) => {
-      app.auction.nextTipCoordinates = returnValue;
-    },
-  });
-
-  auction.steps.push({
-    component: Box,
-    props: {
-      title: 'SSP',
-      x: () => app.auction.nextTipCoordinates?.x - box.width / 2,
-      y: () => app.auction.nextTipCoordinates?.y + config.flow.arrowSize,
-      color: colors.box.noData,
-    },
-    delay: 1000,
-    callBack: (returnValue) => {
-      app.auction.nextTipCoordinates = returnValue.down;
-    },
-  });
-
-  auction.steps.push({
-    component: ProgressLine,
-    props: {
-      direction: 'down',
-      x1: () => app.auction.nextTipCoordinates?.x,
-      y1: () => app.auction.nextTipCoordinates?.y + 40,
-    },
-    callBack: (returnValue) => {
-      app.auction.nextTipCoordinates = returnValue;
-    },
-  });
-
-  auction.steps.push({
-    component: Box,
-    props: {
-      title: 'DSPs',
-      x: () => app.auction.nextTipCoordinates?.x - box.width / 2,
-      y: () => app.auction.nextTipCoordinates?.y + config.flow.arrowSize,
-      color: colors.box.noData,
-    },
-    delay: 1000,
-    callBack: (returnValue) => {
-      app.auction.nextTipCoordinates = returnValue.down;
-    },
-  });
-
-  auction.steps.push({
-    component: ProgressLine,
-    props: {
-      direction: 'up',
-      x1: () => app.auction.nextTipCoordinates?.x + 10,
-      y1: () => {
-        return app.auction.nextTipCoordinates?.y - 15;
-      },
-    },
-    callBack: (returnValue) => {
-      app.auction.nextTipCoordinates = returnValue;
-    },
-  });
-
-  auction.steps.push({
-    component: ProgressLine,
-    props: {
-      direction: 'up',
-      x1: () => app.auction.nextTipCoordinates?.x,
-      y1: () => {
-        return app.auction.nextTipCoordinates?.y - box.height - 10;
-      },
-    },
-    callBack: (returnValue) => {
-      app.auction.nextTipCoordinates = returnValue;
-    },
-  });
-
-  auction.steps.push({
-    component: ProgressLine,
-    props: {
-      direction: 'right',
-      x1: () => app.auction.nextTipCoordinates?.x - 10 + box.width / 2,
-      y1: () => {
-        return app.auction.nextTipCoordinates?.y - 10 - box.height / 2;
-      },
-    },
-    callBack: (returnValue) => {
-      app.auction.nextTipCoordinates = returnValue;
-    },
-  });
-};
-
-auction.setUpMultiSellerFirstSSPTagFlow = () => {
-  const { box, colors } = config.flow;
-
-  auction.steps.push({
-    component: ProgressLine,
-    props: {
-      direction: 'down',
-      x1: () => app.auction.nextTipCoordinates?.x,
-      y1: () => app.auction.nextTipCoordinates?.y + 40,
-      noArrow: true,
-    },
-    callBack: (returnValue) => {
-      app.auction.nextTipCoordinates = returnValue;
-    },
-  });
-
-  auction.steps.push({
-    component: Box,
-    props: {
-      title: 'SSP Adapter',
-      description: 'header-bidding lib',
-      x: () => app.auction.nextTipCoordinates?.x - box.width / 2,
-      y: () => app.auction.nextTipCoordinates?.y,
-    },
-    delay: 1000,
-    callBack: (returnValue) => {
-      app.auction.nextTipCoordinates = returnValue.down;
-    },
-  });
-
-  auction.steps.push({
-    component: ProgressLine,
-    props: {
-      direction: 'down',
-      x1: () => app.auction.nextTipCoordinates?.x,
-      y1: () => app.auction.nextTipCoordinates?.y + 40,
-    },
-    callBack: (returnValue) => {
-      app.auction.nextTipCoordinates = returnValue;
-    },
-  });
-
-  auction.steps.push({
-    component: Box,
-    props: {
-      title: 'SSPs',
-      x: () => app.auction.nextTipCoordinates?.x - box.width / 2,
-      y: () => app.auction.nextTipCoordinates?.y + config.flow.arrowSize,
-      color: colors.box.noData,
-    },
-    delay: 1000,
-    callBack: (returnValue) => {
-      app.auction.nextTipCoordinates = returnValue.down;
-    },
-  });
-
-  auction.steps.push({
-    component: ProgressLine,
-    props: {
-      direction: 'down',
-      x1: () => app.auction.nextTipCoordinates?.x,
-      y1: () => app.auction.nextTipCoordinates?.y + 40,
-    },
-    callBack: (returnValue) => {
-      app.auction.nextTipCoordinates = returnValue;
-    },
-  });
-
-  auction.steps.push({
-    component: Box,
-    props: {
-      title: 'DSPs',
-      x: () => app.auction.nextTipCoordinates?.x - box.width / 2,
-      y: () => app.auction.nextTipCoordinates?.y + config.flow.arrowSize,
-      color: colors.box.noData,
-    },
-    delay: 1000,
-    callBack: (returnValue) => {
-      app.auction.nextTipCoordinates = returnValue.down;
-    },
-  });
-
-  auction.steps.push({
-    component: ProgressLine,
-    props: {
-      direction: 'up',
-      x1: () => app.auction.nextTipCoordinates?.x + 10,
-      y1: () => {
-        return app.auction.nextTipCoordinates?.y - 15;
-      },
-    },
-    callBack: (returnValue) => {
-      app.auction.nextTipCoordinates = returnValue;
-    },
-  });
-
-  auction.steps.push({
-    component: ProgressLine,
-    props: {
-      direction: 'up',
-      x1: () => app.auction.nextTipCoordinates?.x,
-      y1: () => {
-        return app.auction.nextTipCoordinates?.y - box.height - 10;
-      },
-    },
-    callBack: (returnValue) => {
-      app.auction.nextTipCoordinates = returnValue;
-    },
-  });
-
-  auction.steps.push({
-    component: ProgressLine,
-    props: {
-      direction: 'right',
-      x1: () => app.auction.nextTipCoordinates?.x - 10 + box.width / 2,
-      y1: () => {
-        return app.auction.nextTipCoordinates?.y - 10 - box.height / 2;
-      },
-    },
-    callBack: (returnValue) => {
-      app.auction.nextTipCoordinates = returnValue;
-    },
-  });
-};
-
-auction.setUpPublisherAdServerFlow = () => {
-  const { box } = config.flow;
-
-  auction.steps.push({
-    component: Box,
-    props: {
-      title: 'Publisher',
-      description: 'Ad server tag',
-      x: () => app.auction.nextTipCoordinates?.x + 10,
-      y: () => app.auction.nextTipCoordinates?.y - box.height / 2,
-    },
-    delay: 1000,
-    callBack: (returnValue) => {
-      app.auction.nextTipCoordinates = returnValue.down;
-    },
-  });
-
-  auction.steps.push({
-    component: ProgressLine,
-    props: {
-      direction: 'right',
-      x1: () => app.auction.nextTipCoordinates?.x + box.width / 2,
-      y1: () => app.auction.nextTipCoordinates?.y + 10,
-    },
-    callBack: (returnValue) => {
-      app.auction.nextTipCoordinates = returnValue;
-    },
-  });
-
-  auction.steps.push({
-    component: Box,
-    props: {
-      title: 'Publisher',
-      description: 'Ad server',
-      x: () => app.auction.nextTipCoordinates?.x + 10,
-      y: () => app.auction.nextTipCoordinates?.y - 23,
-    },
-    delay: 1000,
-    callBack: (returnValue) => {
-      app.auction.nextTipCoordinates = returnValue.down;
-    },
-  });
-
-  auction.steps.push({
-    component: ProgressLine,
-    props: {
-      direction: 'left',
-      x1: () => app.auction.nextTipCoordinates?.x + box.width / 4 + 6,
-      y1: () => app.auction.nextTipCoordinates?.y + 10,
-    },
-    callBack: (returnValue) => {
-      app.auction.nextTipCoordinates = returnValue;
-    },
-  });
-
-  auction.steps.push({
-    component: ProgressLine,
-    props: {
-      direction: 'down',
-      x1: () => app.auction.nextTipCoordinates?.x - box.width / 2 - 10,
-      y1: () => app.auction.nextTipCoordinates?.y + box.height / 2 - 5,
-    },
-    callBack: (returnValue) => {
-      app.auction.nextTipCoordinates = returnValue;
-    },
-  });
-
-  auction.steps.push({
-    component: Box,
-    props: {
-      title: 'runAdAuction',
-      x: () => app.auction.nextTipCoordinates?.x - box.width / 2,
-      y: () => app.auction.nextTipCoordinates?.y + box.height / 4,
-    },
-    delay: 1000,
-    callBack: (returnValue) => {
-      app.auction.nextTipCoordinates = returnValue.down;
-    },
-  });
-};
-
-auction.componentAuction = () => {
-  const { box } = config.flow;
-
-  const BORDER_WIDTH = 400;
-  const BORDER_HEIGHT = BORDER_WIDTH;
-  const BORDER_BOX_MARGIN = 50;
-
-  auction.steps.push({
-    component: ProgressLine,
-    props: {
-      direction: 'down',
-      x1: () => app.auction.nextTipCoordinates?.x,
-      y1: () => app.auction.nextTipCoordinates?.y + box.height - 10,
-      customHeight: 150,
-      noArrow: true,
-    },
-    callBack: (returnValue) => {
-      app.auction.coordinates.temp = returnValue;
-      app.auction.nextTipCoordinates = returnValue;
-    },
-  });
-
-  auction.steps.push({
-    component: ProgressLine,
-    props: {
-      direction: 'right',
-      x1: () => app.auction.nextTipCoordinates?.x,
-      y1: () => app.auction.nextTipCoordinates?.y,
-      customWidth: BORDER_WIDTH,
-      noArrow: true,
-    },
-    callBack: () => {
-      app.auction.nextTipCoordinates = app.auction.coordinates.temp;
-    },
-  });
-
-  auction.steps.push({
-    component: ProgressLine,
-    props: {
-      direction: 'left',
-      x1: () => app.auction.nextTipCoordinates?.x + 150,
-      y1: () => app.auction.nextTipCoordinates?.y - 10,
-      customWidth: BORDER_WIDTH,
-      noArrow: true,
-    },
-    callBack: (returnValue) => {
-      app.auction.nextTipCoordinates = returnValue;
-    },
-  });
-
-  auction.steps.push({
-    component: ProgressLine,
-    props: {
-      direction: 'down',
-      x1: () => app.auction.nextTipCoordinates?.x,
-      y1: () => app.auction.nextTipCoordinates?.y,
-      customHeight: BORDER_HEIGHT,
-      noArrow: true,
-    },
-    callBack: (returnValue) => {
-      app.auction.nextTipCoordinates = returnValue;
-    },
-  });
-
-  auction.steps.push({
-    component: Box,
-    props: {
-      title: 'SSP A',
-      x: () => app.auction.nextTipCoordinates?.x + BORDER_BOX_MARGIN,
-      y: () =>
-        app.auction.nextTipCoordinates?.y - BORDER_HEIGHT + BORDER_BOX_MARGIN,
-    },
-    delay: 1000,
-    callBack: (returnValue) => {
-      const val = returnValue.down;
-
-      val.x = val.x - box.width / 2 - 10;
-      val.y = val.y + box.height * 2;
-
-      app.auction.nextTipCoordinates = val;
-    },
-  });
-
-  auction.steps.push({
-    component: ProgressLine,
-    props: {
-      direction: 'down',
-      x1: () => app.auction.nextTipCoordinates?.x,
-      y1: () => app.auction.nextTipCoordinates?.y,
-      customHeight: BORDER_HEIGHT,
-      noArrow: true,
-    },
-    callBack: (returnValue) => {
-      app.auction.nextTipCoordinates = returnValue;
-    },
-  });
-
-  auction.setUpRunadAuction();
-
-  auction.steps.push({
-    component: ProgressLine,
-    props: {
-      direction: 'down',
-      x1: () => app.auction.nextTipCoordinates?.x,
-      y1: () => app.auction.nextTipCoordinates?.y,
-      customHeight: BORDER_HEIGHT,
-      noArrow: true,
-    },
-    delay: 11111111111,
-    callBack: (returnValue) => {
-      app.auction.nextTipCoordinates = returnValue;
-    },
-  });
-};
-
-auction.setUpRunadAuction = () => {
-  const steps = auction.steps;
-  const { box, colors } = config.flow;
-
-  steps.push({
-    component: Box,
-    props: {
-      title: 'runAdAuction',
-      x: () => app.auction.nextTipCoordinates?.x + 10,
-      y: () => app.auction.nextTipCoordinates?.y - box.height / 2,
-    },
-    delay: 1000,
-    callBack: (returnValue) => {
-      app.auction.nextTipCoordinates = returnValue.down;
-    },
-  });
-
-  const boxes = [
-    {
-      title: 'Load Interest Group',
-      extraProps: {
-        showBarrageAnimation: true,
-      },
-    },
-    {
-      title: 'Key/Value Trusted',
-      description: 'DSP Server',
-      color: colors.box.notBrowser,
-    },
-    {
-      title: 'generateBid()',
-      description: '(from DSPs on dsp.js)',
-      color: colors.box.notBrowser,
-      extraProps: {
-        showRippleEffect: true,
-      },
-    },
-    {
-      title: 'DSP 1',
-    },
-    {
-      title: 'DSP 2',
-    },
-    {
-      title: 'Key/Value Trusted',
-      description: 'SSP Server',
-    },
-    {
-      title: 'scoreAd()',
-      description: '(by SSPs on ssp.js)',
-    },
-    {
-      title: 'reportWin()',
-      description: '(on dsp.js)',
-      color: colors.box.notBrowser,
-    },
-    {
-      title: 'reportResult()',
-      description: '(on ssp.js)',
-    },
-  ];
-
-  boxes.forEach(({ title, description, color, extraProps = {} }) => {
-    if (title === 'DSP 1') {
-      steps.push({
-        component: ProgressLine,
-        props: {
-          direction: 'right',
-          x1: () => app.auction.nextTipCoordinates?.x + box.width / 2,
-          y1: () => {
-            return app.auction.nextTipCoordinates?.y - box.height / 2 + 15;
-          },
-        },
-        callBack: (returnValue) => {
-          app.auction.nextTipCoordinates = returnValue;
-        },
-      });
-
-      steps.push({
-        component: Box,
-        props: {
-          title,
-          x: () => app.auction.nextTipCoordinates?.x + 10,
-          y: () => app.auction.nextTipCoordinates?.y - box.height + 15,
-          color,
-        },
-        delay: 1000,
-        callBack: (returnValue) => {
-          app.auction.nextTipCoordinates = returnValue.down;
-        },
-      });
-
-      steps.push({
-        component: ProgressLine,
-        props: {
-          direction: 'left',
-          text: '$20',
-          x1: () => app.auction.nextTipCoordinates?.x + box.width / 4 + 5,
-          y1: () => {
-            return app.auction.nextTipCoordinates?.y + 25;
-          },
-        },
-        callBack: (returnValue) => {
-          app.auction.nextTipCoordinates = returnValue;
-        },
-      });
-      return;
-    }
-
-    if (title === 'DSP 2') {
-      steps.push({
-        component: ProgressLine,
-        props: {
-          direction: 'right',
-          x1: () => app.auction.nextTipCoordinates?.x - 7,
-          y1: () => {
-            return app.auction.nextTipCoordinates?.y + box.height / 2 - 5;
-          },
-        },
-        callBack: (returnValue) => {
-          app.auction.nextTipCoordinates = returnValue;
-        },
-      });
-
-      steps.push({
-        component: Box,
-        props: {
-          title,
-          x: () => app.auction.nextTipCoordinates?.x + 10,
-          y: () => app.auction.nextTipCoordinates?.y,
-          color,
-        },
-        delay: 1000,
-        callBack: (returnValue) => {
-          app.auction.nextTipCoordinates = returnValue.down;
-        },
-      });
-
-      steps.push({
-        component: ProgressLine,
-        props: {
-          direction: 'left',
-          text: '$10',
-          x1: () => app.auction.nextTipCoordinates?.x + box.width / 4 + 5,
-          y1: () => {
-            return app.auction.nextTipCoordinates?.y + 15 - box.height / 2;
-          },
-        },
-        callBack: (returnValue) => {
-          app.auction.nextTipCoordinates = {
-            x: returnValue.x - box.width / 2 - 10,
-            y: returnValue.y - (box.height / 4) * 3,
-          };
-        },
-      });
-      return;
-    }
-
-    steps.push({
-      component: ProgressLine,
-      props: {
-        direction: 'down',
-        x1: () => app.auction.nextTipCoordinates?.x,
-        y1: () => {
-          return app.auction.nextTipCoordinates?.y + box.height - 10;
-        },
-      },
-      callBack: (returnValue) => {
-        app.auction.nextTipCoordinates = returnValue;
-      },
-    });
-
-    steps.push({
-      component: Box,
-      props: {
-        title,
-        description,
-        ...extraProps,
-        x: () => app.auction.nextTipCoordinates?.x - box.width / 2,
-        y: () => app.auction.nextTipCoordinates?.y + 10,
-        color,
-      },
-      delay: 1000,
-      callBack: (returnValue) => {
-        app.auction.nextTipCoordinates = returnValue.down;
-      },
-    });
   });
 };
